# docs/openapi.yaml
openapi: 3.0.0
info:
  title: MyTrello API
  version: 0.1.9
  contact:
    name: Shagalov Veaceslav
    url: t.me/slava_shagalov
servers:
  - url: http://localhost/api/v1
    description: Main server
...
paths:
...
  /workspaces:
    post:
      tags:
        - Workspaces
      summary: Create a new workspace
      description: Create a new workspace
      requestBody:
        description: A JSON object containing the data to create workspace.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspace"
      responses:
        "200":
          description: Created workspace data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    get:
      tags:
        - Workspaces
      summary: Returns all workspaces with boards of current user
      description: Returns all workspaces with boards of current user
      responses:
        "200":
          description: A JSON with workspaces array
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspacesWithBoards"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
  /workspaces/{id}:
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
    get:
      tags:
        - Workspaces
      summary: Returns workspace by id
      description: Returns workspace by id
      responses:
        "200":
          description: Workspace data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrWorkspaceNotFound"
        "500":
          $ref: "#/components/responses/ErrInternalServerError"
      security:
        - cookieAuth: [ ]
    patch:
      tags:
        - Workspaces
      summary: Partial update of workspace
      description: Partial update of workspace
      requestBody:
        description: A JSON object containing the data to update workspace.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartialUpdateWorkspace"
      responses:
        "200":
          description: A JSON object of updated workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "403":
          $ref: "#/components/responses/ErrForbidden"
        "404":
          $ref: "#/components/responses/ErrWorkspaceNotFound"
      security:
        - cookieAuth: [ ]
    delete:
      tags:
        - Workspaces
      summary: Delete workspace by id
      description: Delete workspace by id
      responses:
        "204":
          description: Workspace deleted successfully.
        "400":
          $ref: "#/components/responses/ErrResponseBadRequest"
        "401":
          $ref: "#/components/responses/ErrUnauthorized"
        "404":
          $ref: "#/components/responses/ErrWorkspaceNotFound"
      security:
        - cookieAuth: [ ]
...
components:
  schemas:
    Id:
      type: integer
      required: true
      example: 3
...
    MediaSource:
      type: string
      example: https://s3.trello.ru/uploaded_picture.jpg
    Time:
      type: string
      example: "11.07.2023 12:23:45"
...
    Workspace:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "IT"
        description:
          type: string
          example: "Information Technologies"
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
    WorkspaceWithBoards:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "IT"
        description:
          type: string
          example: "Information Technologies"
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
        boards:
          type: array
          items:
            $ref: "#/components/schemas/Board"
...
    WorkspacesWithBoards:
      type: object
      properties:
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceWithBoards"
    CreateWorkspace:
      type: object
      properties:
        title:
          type: string
          required: true
        description:
          type: string
          required: false
      example: {
        "title": "IT",
        "description": "Information Technologies",
      }
    PartialUpdateWorkspace:
      type: object
      properties:
        title:
          type: string
          required: false
          example: "IT"
        description:
          type: string
          required: false
          example: "Information Technologies"
    Board:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/Id"
        workspace_id:
          $ref: "#/components/schemas/Id"
        title:
          type: string
          example: "Backend"
        description:
          type: string
          example: "Backend Board"
        background:
          $ref: "#/components/schemas/MediaSource"
        created_at:
          $ref: "#/components/schemas/Time"
        updated_at:
          $ref: "#/components/schemas/Time"
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    ErrResponseBadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            description: Bad request
            properties:
              error:
                type: string
                example: "{{reason}}"
    ErrUnauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            description: Unauthorized
            properties:
              error:
                type: string
                example: "unauthorized"
...
    ErrWorkspaceNotFound:
      description: Workspace not found
      content:
        application/json:
          schema:
            type: object
            description: Workspace not found
            properties:
              error:
                type: string
                example: "workspace not found"
...
    ErrForbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            description: Access denied
            properties:
              error:
                type: string
                example: "access denied"
    ErrInternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            description: Internal Server Error
            properties:
              error:
                type: string
                example: "{{reason}}"